+++ lib/actions/register.ts
@@ -4,6 +4,60 @@ import { z } from "zod";
 import { prisma } from "@/prisma/prisma";
 import bcrypt from 'bcrypt';
 import crypto from 'crypto';
+import nodemailer from 'nodemailer';
+
+// Create a nodemailer transporter with Gmail SMTP
+const transporter = nodemailer.createTransport({
+    service: 'gmail',
+    auth: {
+        user: process.env.GMAIL_EMAIL_ADDRESS,
+        pass: process.env.GMAIL_APP_PASSWORD,
+    },
+});
+
+// Create HTML template for verification email
+function createVerificationEmailHTML(name: string, verificationUrl: string) {
+    return `
+    <div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto; padding: 20px; background-color: #f9f9f9; border-radius: 8px;">
+      <h1 style="color: #333; text-align: center;">Welcome to Employment Opportunities!</h1>
+      <p style="font-size: 16px; line-height: 1.5; color: #555;">Hello ${name},</p>
+      <p style="font-size: 16px; line-height: 1.5; color: #555;">
+        Thank you for registering. Please click the button below to verify your email address:
+      </p>
+      <div style="text-align: center; margin: 30px 0;">
+        <a 
+          href="${verificationUrl}"
+          style="background-color: #22c55e; color: white; padding: 12px 24px; border-radius: 4px; text-decoration: none; font-weight: bold; display: inline-block;"
+        >
+          Verify my email
+        </a>
+      </div>
+      <p style="font-size: 16px; line-height: 1.5; color: #555;">
+        This link will expire in 24 hours.
+      </p>
+      <p style="font-size: 14px; color: #777; margin-top: 40px; border-top: 1px solid #eee; padding-top: 20px;">
+        If you didn't create this account, you can safely ignore this email.
+      </p>
+    </div>
+  `;
+}
+
+// Create plain text version for email clients that don't support HTML
+function createVerificationEmailText(name: string, verificationUrl: string) {
+    return `
+    Welcome to Employment Opportunities!
+    
+    Hello ${name},
+    
+    Thank you for registering. Please visit the following link to verify your email address:
+    
+    ${verificationUrl}
+    
+    This link will expire in 24 hours.
+    
+    If you didn't create this account, you can safely ignore this email.
+  `;
+}
 
 // Schema for validation
 const SignupSchema = z.object({
@@ -80,24 +134,24 @@ export async function signUp(formData: FormData | { name: string; email: string;
         });
 
         // Base URL for verification link
-        const baseUrl = process.env.NEXTAUTH_URL || 'http://localhost:3000';
+        const baseUrl = process.env.NEXTAUTH_URL || 'https://employment-opportunities.vercel.app';
         const verificationUrl = `${baseUrl}/api/auth/verify?token=${token}&email=${encodeURIComponent(email)}`;
 
-        // Send verification email using the dedicated API route
-        const emailResponse = await fetch(`${baseUrl}/api/email/send-verification`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-                email,
-                name,
-                verificationUrl
-            }),
-        });
+        try {
+            // Send the verification email directly instead of using the API route
+            const mailOptions = {
+                from: `"Employment Opportunities" <${process.env.GMAIL_EMAIL_ADDRESS}>`,
+                to: email,
+                subject: 'Verify your email address',
+                text: createVerificationEmailText(name, verificationUrl),
+                html: createVerificationEmailHTML(name, verificationUrl),
+            };
 
-        if (!emailResponse.ok) {
-            console.error('Failed to send verification email:', await emailResponse.text());
+            await transporter.sendMail(mailOptions);
+        } catch (emailError) {
+            console.error('Failed to send verification email:', emailError);
+            // Don't fail the signup if email sending fails
+            // The token is already in the database and the user can still sign up
         }
 
         // Return success response (excluding password)
